cleanup + docs
